upgrade: Add logic for 0.11->1.0 upgrade
This also packages operations a little bit more efficiently. Also, in 1.0, the whole Housekeeping
logic is a little bit different.
