+++ lib/upgrade/0.10_to_0.11.go
@@ -21,17 +21,9 @@ package upgrade
 import (
 	"context"
 	"fmt"
-	"time"
 
-	"github.com/astarte-platform/astarte-kubernetes-operator/lib/reconcile"
 	apiv1alpha1 "github.com/astarte-platform/astarte-kubernetes-operator/pkg/apis/api/v1alpha1"
-	"github.com/astarte-platform/astarte-kubernetes-operator/pkg/misc"
-	"github.com/openlyinc/pointy"
-	appsv1 "k8s.io/api/apps/v1"
-	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -65,7 +57,7 @@ func upgradeTo011(cr *apiv1alpha1.Astarte, c client.Client, scheme *runtime.Sche
 	recorder.Event(cr, "Normal", apiv1alpha1.AstarteResourceEventUpgrade.String(),
 		"Starting Database Migration")
 	reqLogger.Info("Upgrading Housekeeping and migrating the Database...")
-	housekeepingBackend, err := upgradeHousekeeping(cr, c, scheme, recorder)
+	housekeepingBackend, err := upgradeHousekeeping(landing011Version, true, cr, c, scheme, recorder)
 	if err != nil {
 		return err
 	}
@@ -89,7 +81,7 @@ func upgradeTo011(cr *apiv1alpha1.Astarte, c client.Client, scheme *runtime.Sche
 	reqLogger.Info("Ensuring new RabbitMQ Queue Layout through Data Updater Plant...")
 	recorder.Event(cr, "Normal", apiv1alpha1.AstarteResourceEventUpgrade.String(),
 		"Starting Queue layout migration")
-	if err := waitForQueueLayoutMigration(cr, c, scheme); err != nil {
+	if err := waitForQueueLayoutMigration(landing011Version, cr, c, scheme); err != nil {
 		recorder.Event(cr, "Warning", apiv1alpha1.AstarteResourceEventCriticalError.String(),
 			"Could not migrate data queues. Upgrade will be retried, but manual intervention is likely required")
 		return fmt.Errorf("Failed in waiting for Data Updater Plant to come up: %v", err)
@@ -112,23 +104,7 @@ func upgradeTo011(cr *apiv1alpha1.Astarte, c client.Client, scheme *runtime.Sche
 	// Just to be sure, scale down Housekeeping to 0 replicas. If we're *really* tight on resources, it might be that
 	// the additional pool prevents other pods from coming up.
 	reqLogger.Info("Restoring original environment and waiting for cluster to settle...")
-	housekeepingBackend.Replicas = pointy.Int32(0)
-	if err := reconcile.EnsureAstarteGenericBackend(cr, *housekeepingBackend, apiv1alpha1.Housekeeping, c, scheme); err != nil {
-		return err
-	}
-	// Wait for it to go down, then we should be good to go.
-	if err := wait.Poll(retryInterval, timeout, func() (done bool, err error) {
-		deployment := &appsv1.Deployment{}
-		if err = c.Get(context.TODO(), types.NamespacedName{Name: cr.Name + "-housekeeping", Namespace: cr.Namespace}, deployment); err != nil {
-			return false, err
-		}
-
-		if deployment.Status.ReadyReplicas > 0 {
-			return false, nil
-		}
-
-		return true, nil
-	}); err != nil {
+	if err := scaleDownHousekeeping(housekeepingBackend, cr, c, scheme); err != nil {
 		reqLogger.Error(err, "Failed in waiting for Housekeeping to go down. Continuing anyway.")
 	}
 
@@ -136,148 +112,4 @@ func upgradeTo011(cr *apiv1alpha1.Astarte, c client.Client, scheme *runtime.Sche
 	recorder.Event(cr, "Normal", apiv1alpha1.AstarteResourceEventUpgrade.String(),
 		"Astarte upgraded successfully to the 0.11.x series")
 	return nil
-}
-
-func waitForQueueLayoutMigration(cr *apiv1alpha1.Astarte, c client.Client, scheme *runtime.Scheme) error {
-	dataUpdaterPlant := cr.Spec.Components.DataUpdaterPlant.DeepCopy()
-	dataUpdaterPlant.Version = landing011Version
-	// Ensure the policy is Replace. We don't want to have old pods hanging around.
-	dataUpdaterPlant.DeploymentStrategy = &appsv1.DeploymentStrategy{Type: appsv1.RecreateDeploymentStrategyType}
-	if err := reconcile.EnsureAstarteGenericBackend(cr, dataUpdaterPlant.AstarteGenericClusteredResource, apiv1alpha1.DataUpdaterPlant, c, scheme); err != nil {
-		return err
-	}
-	// The operation should be pretty normal and quick enough. Wait with standard timeouts here
-	return wait.Poll(retryInterval, timeout, func() (done bool, err error) {
-		deployment := &appsv1.Deployment{}
-		if err = c.Get(context.TODO(), types.NamespacedName{Name: cr.Name + "-data-updater-plant", Namespace: cr.Namespace}, deployment); err != nil {
-			return false, err
-		}
-
-		if deployment.Status.ReadyReplicas > 0 {
-			return true, nil
-		}
-
-		return false, nil
-	})
-}
-
-func waitForHousekeepingUpgrade(cr *apiv1alpha1.Astarte, c client.Client, recorder record.EventRecorder) error {
-	weirdFailuresCount := 0
-	weirdFailuresThreshold := 10
-
-	return wait.Poll(retryInterval, time.Hour, func() (done bool, err error) {
-		deployment := &appsv1.Deployment{}
-		if err = c.Get(context.TODO(), types.NamespacedName{Name: cr.Name + "-housekeeping-api", Namespace: cr.Namespace}, deployment); err != nil {
-			weirdFailuresCount++
-			if weirdFailuresCount > weirdFailuresThreshold {
-				// Something is off.
-				recorder.Event(cr, "Warning", apiv1alpha1.AstarteResourceEventCriticalError.String(),
-					"Repeated errors in monitoring Database Migration. Manual intervention is likely required")
-				return false, fmt.Errorf("Failed in looking up Housekeeping API Deployment. Most likely, manual intervention is required. %v", err)
-			}
-			// Something is off.
-			log.Error(err, "Failed in looking up Housekeeping API Deployment. This might be a temporary problem - will retry")
-			return false, nil
-		}
-
-		if deployment.Status.ReadyReplicas >= 1 {
-			// That's it bros.
-			return true, nil
-		}
-
-		// Ensure we aren't in the position where Housekeeping itself is crashing.
-		housekeepingComponent := apiv1alpha1.Housekeeping
-		podList := &v1.PodList{}
-		if err = c.List(context.TODO(), podList, client.InNamespace(cr.Namespace),
-			client.MatchingLabels{"astarte-component": housekeepingComponent.DashedString()}); err != nil {
-			weirdFailuresCount++
-			if weirdFailuresCount > weirdFailuresThreshold {
-				// Something is off.
-				recorder.Event(cr, "Warning", apiv1alpha1.AstarteResourceEventCriticalError.String(),
-					"Repeated errors in monitoring Database Migration. Manual intervention is likely required")
-				return false, fmt.Errorf("Failed in looking up Housekeeping pods. Most likely, manual intervention is required. %v", err)
-			}
-			// Something is off.
-			log.Error(err, "Failed in looking up Housekeeping pods. This might be a temporary problem - will retry")
-			return false, nil
-		}
-
-		// Inspect the list!
-		if len(podList.Items) != 1 {
-			weirdFailuresCount++
-			if weirdFailuresCount > weirdFailuresThreshold {
-				// Something is off.
-				recorder.Event(cr, "Warning", apiv1alpha1.AstarteResourceEventCriticalError.String(),
-					"Repeated errors in monitoring Database Migration. Manual intervention is likely required")
-				return false, fmt.Errorf("%v Housekeeping pods found. Most likely, manual intervention is required. %v", len(podList.Items), err)
-			}
-			// Something is off.
-			log.Error(err, fmt.Sprintf("%v Housekeeping pods found. This might be a temporary problem - will retry", len(podList.Items)))
-			return false, nil
-		}
-
-		if len(podList.Items[0].Status.ContainerStatuses) != 1 {
-			weirdFailuresCount++
-			if weirdFailuresCount > weirdFailuresThreshold {
-				// Something is off.
-				recorder.Event(cr, "Warning", apiv1alpha1.AstarteResourceEventCriticalError.String(),
-					"Repeated errors in monitoring Database Migration. Manual intervention is likely required")
-				return false, fmt.Errorf("%v Container Statuses retrieved. Most likely, manual intervention is required. %v", len(podList.Items[0].Status.ContainerStatuses), err)
-			}
-			// Something is off.
-			log.Error(err, fmt.Sprintf("%v Container Statuses retrieved. This might be a temporary problem - will retry", len(podList.Items[0].Status.ContainerStatuses)))
-			return false, nil
-		}
-
-		if podList.Items[0].Status.ContainerStatuses[0].State.Waiting != nil {
-			if podList.Items[0].Status.ContainerStatuses[0].State.Waiting.Reason == "CrashLoopBackoff" {
-				recorder.Event(cr, "Warning", apiv1alpha1.AstarteResourceEventCriticalError.String(),
-					"Database Migration failed. Manual intervention is likely required")
-				return true, fmt.Errorf("Housekeeping is crashing repeatedly. There has to be a problem in handling Database migrations. Please take manual action as soon as possible")
-			}
-		}
-
-		return false, nil
-	})
-}
-
-func upgradeHousekeeping(cr *apiv1alpha1.Astarte, c client.Client, scheme *runtime.Scheme,
-	recorder record.EventRecorder) (*apiv1alpha1.AstarteGenericClusteredResource, error) {
-	housekeepingBackend := cr.Spec.Components.Housekeeping.Backend.DeepCopy()
-	housekeepingBackend.Version = landing011Version
-	housekeepingBackend.Replicas = pointy.Int32(1)
-	// Ensure the policy is Replace. We don't want to have old pods hanging around.
-	housekeepingBackend.DeploymentStrategy = &appsv1.DeploymentStrategy{Type: appsv1.RecreateDeploymentStrategyType}
-	if cr.Spec.VerneMQ.Resources != nil {
-		resourceRequirements := misc.GetResourcesForAstarteComponent(cr, housekeepingBackend.Resources, apiv1alpha1.Housekeeping)
-		resourceRequirements.Requests.Cpu().Add(*cr.Spec.VerneMQ.Resources.Requests.Cpu())
-		resourceRequirements.Requests.Memory().Add(*cr.Spec.VerneMQ.Resources.Requests.Memory())
-		resourceRequirements.Limits.Cpu().Add(*cr.Spec.VerneMQ.Resources.Limits.Cpu())
-		resourceRequirements.Limits.Memory().Add(*cr.Spec.VerneMQ.Resources.Limits.Memory())
-
-		// This way, on the next call to GetResourcesForAstarteComponent, these resources will be returned as explicitly stated
-		// in the original spec.
-		housekeepingBackend.Resources = &resourceRequirements
-	}
-
-	// Add a custom, more permissive probe to the Backend
-	if err := reconcile.EnsureAstarteGenericBackendWithCustomProbe(cr, *housekeepingBackend, apiv1alpha1.Housekeeping,
-		c, scheme, getSpecialHousekeepingMigrationProbe("/health")); err != nil {
-		recorder.Event(cr, "Warning", apiv1alpha1.AstarteResourceEventUpgradeError.String(),
-			"Could not initiate Database Migration. Upgrade will be retried")
-		return nil, err
-	}
-	housekeepingAPI := cr.Spec.Components.Housekeeping.API.DeepCopy()
-	housekeepingAPI.Replicas = pointy.Int32(1)
-	housekeepingAPI.Version = landing011Version
-	// Ensure the policy is Replace. We don't want to have old pods hanging around.
-	housekeepingAPI.DeploymentStrategy = &appsv1.DeploymentStrategy{Type: appsv1.RecreateDeploymentStrategyType}
-	if err := reconcile.EnsureAstarteGenericAPIWithCustomProbe(cr, *housekeepingAPI, apiv1alpha1.HousekeepingAPI, c,
-		scheme, getSpecialHousekeepingMigrationProbe("/health")); err != nil {
-		recorder.Event(cr, "Warning", apiv1alpha1.AstarteResourceEventUpgradeError.String(),
-			"Could not initiate Database Migration. Upgrade will be retried")
-		return nil, err
-	}
-
-	return housekeepingBackend, nil
 }
